Use one point to check token
