+++ src/Pages/Profile/EditProfile/index.tsx
@@ -1,20 +1,10 @@
-import React, { ReactElement, useEffect } from 'react';
-import { useSelector } from 'react-redux';
-import { useHistory } from 'react-router-dom';
+import React, { ReactElement } from 'react';
 
 import PageHeaderWithoutButton from '../../../Components/Header/PageHeaderWithoutBtn';
-import { selectToken } from '../../../store/user/selectors';
 import PageContent from '../../Sections/PageContent';
 import EditProfileForm from './EditProfileForm';
 
 const EditProfile: React.FC = (): ReactElement => {
-  const token = useSelector(selectToken);
-  const history = useHistory();
-
-  useEffect(() => {
-    if (!token) history.push('/login');
-  });
-
   return (
     <PageContent
       loadingText="Profiel"
