+++ src/Pages/Rules/index.tsx
@@ -1,25 +1,20 @@
-import { Box } from '@material-ui/core';
-import React, { ReactElement, useEffect } from 'react';
-import { useSelector } from 'react-redux';
-import { useHistory } from 'react-router-dom';
+import React, { ReactElement } from 'react';
 
 import PageHeaderWithoutButton from '../../Components/Header/PageHeaderWithoutBtn';
-import { selectToken } from '../../store/user/selectors';
+import PageContent from '../Sections/PageContent';
 import AccordionWithRules from './AccordionWithRules';
 
 const Rules: React.FC = (): ReactElement => {
-  const history = useHistory();
-  const token = useSelector(selectToken);
-
-  useEffect(() => {
-    if (!token) history.push('/login');
-  });
-
   return (
-    <Box>
-      <PageHeaderWithoutButton title="Regels" />
-      <AccordionWithRules />
-    </Box>
+    <PageContent
+      loadingText="Regels"
+      content={
+        <>
+          <PageHeaderWithoutButton title="Regels" />
+          <AccordionWithRules />
+        </>
+      }
+    />
   );
 };
