+++ src/Pages/Scores/TotoRound/index.tsx
@@ -9,7 +9,6 @@ import SubTitleComponent from '../../../Components/Title/SubTitle';
 import { fetchScoresTotoRound } from '../../../store/scores/actions';
 import { selectTotoRound } from '../../../store/scores/selectors';
 import { UserWithScore } from '../../../store/scores/types';
-import { selectToken } from '../../../store/user/selectors';
 import ScoresBarChart from '../../Sections/Charts/ScoresBarChart';
 import PageContent from '../../Sections/PageContent';
 import BreadCrumbsSection from './BreadCrumbsSection';
@@ -17,13 +16,8 @@ import BreadCrumbsSection from './BreadCrumbsSection';
 const TotoRound: React.FC = (): ReactElement => {
   const dispatch = useDispatch();
   const history = useHistory();
-  const token = useSelector(selectToken);
   const totoRound = useSelector(selectTotoRound);
   const { id } = useParams<{ id: string }>();
-
-  useEffect(() => {
-    if (!token) history.push('/login');
-  });
 
   useEffect(() => {
     if (!totoRound || (totoRound && +id !== +totoRound.id)) {
