+++ src/Pages/Sections/PageContent/index.tsx
@@ -4,6 +4,8 @@ import { useSelector } from 'react-redux';
 
 import ProgressComponent from '../../../Components/Progress';
 import { selectAppLoading } from '../../../store/appState/selectors';
+import { selectToken } from '../../../store/user/selectors';
+import CheckLoginStatus from '../../../utils/Auth/checkLoginStatus';
 
 interface IProps {
   loadingText: string;
@@ -11,9 +13,12 @@ interface IProps {
 }
 
 const PageContent: React.FC<IProps> = ({ loadingText, content }: IProps): ReactElement => {
+  const token = useSelector(selectToken);
   const isLoading = useSelector(selectAppLoading);
 
-  return <Box>{isLoading ? <ProgressComponent loadingText={loadingText} /> : content}</Box>;
+  return (
+    <Box>{isLoading ? <ProgressComponent loadingText={loadingText} /> : token ? content : <CheckLoginStatus />}</Box>
+  );
 };
 
 export default PageContent;
